requirements.txt
watchdog==3.0.0
boto3
python-dotenv==1.0.0